Added tons of comments
