+++ systems/controllers/kinematic_centroidal_mpc/kcmpc_reference_generator.cc
@@ -1,102 +1,6 @@
 #include "kcmpc_reference_generator.h"
 #include "multibody/multibody_utils.h"
-
-
-drake::trajectories::PiecewisePolynomial<double> GenerateComTrajectory(const Eigen::Vector3d& current_com,
-                                                                       const std::vector<Eigen::Vector3d>& vel_ewrt_w,
-                                                                       const std::vector<double>& time_points){
-  DRAKE_DEMAND(vel_ewrt_w.size() == time_points.size());
-  auto n_points = vel_ewrt_w.size();
-
-  std::vector<drake::MatrixX<double>> samples(n_points);
-  samples[0] = current_com;
-  for(int i = 1; i<n_points; i++){
-    samples[i] = samples[i-1] + (time_points[i] - time_points[i-1]) * vel_ewrt_w[i-1];
-  }
-  return drake::trajectories::PiecewisePolynomial<double>::FirstOrderHold(time_points, samples);
-}
-
-drake::trajectories::PiecewisePolynomial<double> GenerateGeneralizedPosTrajectory(const Eigen::VectorXd& nominal_stand,
-                                                                                  const Eigen::Vector3d& base_rt_com_ewrt_w,
-                                                                                  const drake::trajectories::PiecewisePolynomial<double>& com_traj,
-                                                                                  int base_pos_start){
-  auto n_points = com_traj.get_segment_times().size();
-  std::vector<drake::MatrixX<double>> samples(n_points);
-  for(int i = 0; i < n_points; i++){
-    samples[i] = nominal_stand;
-    samples[i].block<3,1>(base_pos_start,0, 3, 1) = com_traj.value(com_traj.get_segment_times()[i]) + base_rt_com_ewrt_w;
-  }
-  return drake::trajectories::PiecewisePolynomial<double>::FirstOrderHold(com_traj.get_segment_times(), samples);
-}
-
-drake::trajectories::PiecewisePolynomial<double> GenerateGeneralizedVelTrajectory(const drake::trajectories::PiecewisePolynomial<double>& com_traj,
-                                                                                  int n_v,
-                                                                                  int base_vel_start){
-  auto n_points = com_traj.get_segment_times().size();
-  std::vector<drake::MatrixX<double>> samples(n_points);
-  for(int i = 0; i < n_points; i++){
-    samples[i] = drake::VectorX<double>::Zero(n_v);
-    samples[i].block<3,1>(base_vel_start,0, 3, 1) = com_traj.derivative().value(com_traj.get_segment_times()[i]);
-  }
-  return drake::trajectories::PiecewisePolynomial<double>::ZeroOrderHold(com_traj.get_segment_times(), samples);
-}
-
-int FindCurrentMode(const Gait& active_gait, double time_now){
-  double phase_now = fmod(time_now/active_gait.period, 1);
-  for(int i = 0; i < active_gait.gait_pattern.size(); i++){
-    const auto& mode = active_gait.gait_pattern[i];
-    if(mode.start_phase <= phase_now && mode.end_phase > phase_now){
-      return i;
-    }
-  }
-  DRAKE_ASSERT(false);
-  return 0;
-}
-
-drake::trajectories::PiecewisePolynomial<double> GenerateModeSequence(const std::vector<Gait>& gait_sequence,
-                                                                      const std::vector<double>& time_points){
-  DRAKE_DEMAND(gait_sequence.size() == time_points.size());
-
-  auto traj = gait_sequence[0].ToTrajectory(time_points[0], time_points[1]);
-  for(int i = 1; i<gait_sequence.size() - 1 ; i++){
-    traj.ConcatenateInTime(gait_sequence[i].ToTrajectory(time_points[i], time_points[i + 1]));
-  }
-  return traj;
-}
-
-drake::trajectories::PiecewisePolynomial<double> Gait::ToTrajectory(double current_time, double end_time) const {
-  std::vector<double> break_points;
-  std::vector<drake::MatrixX<double>> samples;
-
-  // Calculate initial mode index, and phase
-  int current_mode = FindCurrentMode(*this, current_time);
-  double current_phase = fmod(current_time/period, 1);
-
-  // Loop until time is greater than end time
-  while(current_time < end_time){
-    // Add the break point for the current time
-    break_points.push_back(current_time);
-    samples.emplace_back(gait_pattern[current_mode].contact_status.cast<double>());
-
-    // Update time based on how much longer the current mode will last
-    current_time += (gait_pattern[current_mode].end_phase - current_phase) * period;
-    // Update the mode and mod if necessary
-    current_mode = (current_mode + 1) % gait_pattern.size();
-    // The new phase is the start phase of the updated mode
-    current_phase = gait_pattern[current_mode].start_phase;
-  }
-  break_points.push_back(end_time);
-  samples.push_back(samples[samples.size()-1]);
-  return drake::trajectories::PiecewisePolynomial<double>::ZeroOrderHold(break_points, samples);
-}
-void Gait::Is_Valid() const {
-  DRAKE_ASSERT(period > 0);
-  DRAKE_ASSERT(gait_pattern[0].start_phase == 0);
-  DRAKE_ASSERT(gait_pattern[gait_pattern.size()-1].end_phase == 1);
-  for(int i = 0; i < gait_pattern.size() - 1; i++){
-    DRAKE_ASSERT(gait_pattern[i].end_phase == gait_pattern[i+1].start_phase);
-  }
-}
+#include "systems/controllers/kinematic_centroidal_mpc/reference_generation_utils.h"
 
 double TimeNextValue(const drake::trajectories::PiecewisePolynomial<double>& trajectory,
                      double current_time,
