+++ systems/controllers/kinematic_centroidal_mpc/kcmpc_reference_generator.h
@@ -3,37 +3,7 @@
 #include <drake/multibody/plant/multibody_plant.h>
 #include <drake/common/trajectories/piecewise_polynomial.h>
 #include "multibody/kinematic/world_point_evaluator.h"
-
-/*!
- * @brief Struct for a given mode as part of a mode sequence
- */
-struct Mode{
-  double start_phase;
-  double end_phase;
-  drake::VectorX<bool> contact_status; /// Vector describing which contacts are active
-};
-
-/*!
- * @brief Struct for defining a gait consisting of a vector of Modes, and a period for the whole gait.
- */
-struct Gait{
-  double period;
-  std::vector<Mode> gait_pattern; /// Vector of modes. Start phase of the first mode must be 0, end phase of the last mode
-                                  /// must be 1, and no time gaps between start and end of sequential modes
-
-  /*!
-   * @brief converts the gait into a trajectory of of the contact timing
-   * @param current_time time for the trajectory to start
-   * @param end_time time for the trajectory to end
-   * @return the trajectory
-   */
-  drake::trajectories::PiecewisePolynomial<double> ToTrajectory(double current_time, double end_time) const;
-
-  /*!
-   * @brief Checks to make sure the gait is valid, Asserts if not valid
-   */
-  void Is_Valid() const;
-};
+#include "systems/controllers/kinematic_centroidal_mpc/gait.h"
 
 /*!
  * @brief struct consisting of a collection of times and samples used for passing inputs to reference generator
